+++ Emulator/Files/Folder.h
@@ -52,7 +52,7 @@ public:
     //
 
     PETName<16> collectionName() override;
-    u64 collectionCount() const override;
+    isize collectionCount() const override;
     PETName<16> itemName(isize nr) const override;
     u64 itemSize(isize nr) const override;
     u8 readByte(isize nr, u64 pos) const override;
